refactor(recall): split pg_recall.py at its two documented seams (#368)
pg_recall.py stood at 833 lines against the 500-line §4.1 limit, and the
craftsmanship gate blocks any commit touching it — including the five-line
trust-policy wiring #368 needs. Split rather than exempted.

Both seams were already marked in the file by its own section banners, so
this follows the structure the code claimed rather than inventing one:

- pg_recall_assembly.py (287 l.) — the budgeted, slot-filled 3-phase context
  assembly that CONSUMES recall's output. Changes when the prompt budget or
  stage model changes.
- pg_recall_weights.py (116 l.) — intent-adaptive WRRF weight profiles.
  Changes when retrieval is recalibrated.
- pg_recall.py (498 l.) — recall orchestration, and a facade re-exporting
  both so no caller or test had to move (same pattern as
  synaptic_plasticity.py).

Assembly imports recall lazily inside the one function that calls it: the
runtime dependency is genuinely one-directional (assembly -> recall), and a
module-level import would close the cycle the facade creates.

Every value moved unchanged — this re-homed code, it retuned nothing.
Verified by the same 4469-test invocation used before the split, with the
identical count passing.

Also wires UNTRUSTED_ORIGIN_FACTOR into the production recall path; it stays
at the identity (1.0) until the pre-registered sweep reports.
